requirements.txt
flask
python-dotenv
authlib
requests
Authlib
pandas
auth0-python
Flask-Cache
flask-cors
flask-httpauth
flask-login
flask-principal
flask-wtf
gunicorn
itsdangerous
Jinja2
MarkupSafe
psycopg2
pyjwt
requests
sortedcontainers
Werkzeug
wsgiref
wtforms
pytest
cffi
click
flask-talisman
pycparser
cryptography

